+++ Specs/Core/Cartesian4Spec.js
@@ -24,6 +24,35 @@ defineSuite([
         expect(cartesian.w).toEqual(4.0);
     });
 
+    it('fromArray creates a Cartesian4', function() {
+        var cartesian = Cartesian4.fromArray([1.0, 2.0, 3.0, 4.0]);
+        expect(cartesian).toEqual(new Cartesian4(1.0, 2.0, 3.0, 4.0));
+    });
+
+    it('fromArray with an offset creates a Cartesian4', function() {
+        var cartesian = Cartesian4.fromArray([0.0, 1.0, 2.0, 3.0, 4.0, 0.0], 1);
+        expect(cartesian).toEqual(new Cartesian4(1.0, 2.0, 3.0, 4.0));
+    });
+
+    it('fromArray creates a Cartesian4 with a result parameter', function() {
+        var cartesian = new Cartesian4();
+        var result = Cartesian4.fromArray([1.0, 2.0, 3.0, 4.0], 0, cartesian);
+        expect(result).toBe(cartesian);
+        expect(result).toEqual(new Cartesian4(1.0, 2.0, 3.0, 4.0));
+    });
+
+    it('fromArray throws without values', function() {
+        expect(function() {
+            Cartesian4.fromArray();
+        }).toThrow();
+    });
+
+    it('fromArray throws with an invalid offset', function() {
+        expect(function() {
+            Cartesian4.fromArray([0.0, 0.0, 0.0, 0.0], 1);
+        }).toThrow();
+    });
+
     it('clone without a result parameter', function() {
         var cartesian = new Cartesian4(1.0, 2.0, 3.0, 4.0);
         var result = cartesian.clone();
@@ -94,6 +123,23 @@ defineSuite([
     it('magnitude', function() {
         var cartesian = new Cartesian4(3.0, 4.0, 5.0, 6.0);
         expect(cartesian.magnitude()).toEqual(Math.sqrt(86.0));
+    });
+
+    it('distance', function() {
+        var distance = Cartesian4.distance(new Cartesian4(1.0, 0.0, 0.0, 0.0), new Cartesian4(2.0, 0.0, 0.0, 0.0));
+        expect(distance).toEqual(1.0);
+    });
+
+    it('distance throws without left', function() {
+        expect(function() {
+            Cartesian4.distance();
+        }).toThrow();
+    });
+
+    it('distance throws without right', function() {
+        expect(function() {
+            Cartesian4.distance(Cartesian4.UNIT_X);
+        }).toThrow();
     });
 
     it('normalize works without a result parameter', function() {